+++ strahlenwerk/Source/Timeline/TimelineData.h
@@ -48,8 +48,8 @@ class TimelineData {
 		void setSceneDuration(ValueTree scene, var duration);
 		void setSceneShaderSource(ValueTree scene, var shaderSource);
 
-		float getLastSceneEndTime();
-		ValueTree getSceneForTime(const float time);
+		int getLastSceneEndTime();
+		ValueTree getSceneForTime(const int time);
 		ValueTree getCurrentScene();
 		int getNewSceneId();
 
@@ -86,7 +86,7 @@ class TimelineData {
 		ValueTree getSequence(ValueTree uniform, const int nthSequence);
 		bool isSequence(ValueTree sequence);
 		ValueTree addSequence(ValueTree uniform, ValueTree sequence, int position = -1);
-		ValueTree addSequence(ValueTree uniform, float absoluteStart, var duration, var interpolation, int position = -1);
+		ValueTree addSequence(ValueTree uniform, int absoluteStart, var duration, var interpolation, int position = -1);
 		ValueTree addSequenceUnchecked(ValueTree uniform, ValueTree sequence, int position);
 		void removeSequence(ValueTree sequence);
 
@@ -100,9 +100,9 @@ class TimelineData {
 		void setSequenceDuration(ValueTree sequence, var duration);
 		void setSequenceInterpolation(ValueTree sequence, var interpolation);
 
-		void setSequencePropertiesForAbsoluteStart(ValueTree sequence, float absoluteStart);
-		float getAbsoluteStartForSequence(ValueTree sequence);
-		ValueTree getSequenceForTime(ValueTree uniform, const float absoluteTime);
+		void setSequencePropertiesForAbsoluteStart(ValueTree sequence, int absoluteStart);
+		int getAbsoluteStartForSequence(ValueTree sequence);
+		ValueTree getSequenceForTime(ValueTree uniform, const int absoluteTime);
 		ValueTree getCurrentSequence(ValueTree uniform);
 		ValueTree getSequenceParentUniform(ValueTree sequence);
